Add files via upload
This project is a Java program designed to demonstrate Object-Oriented Programming (OOP) concepts through the implementation of a Book class hierarchy. The program defines two types of books: EBook and PrintBook, which extend a common superclass Book. The program then sorts an array of these books and displays their details.